[정유수] 펀딩 리스트 데이터 바인딩 진행중
