+++ src/main/java/com/tenco/indiepicter/banner/BannerService.java
@@ -1,6 +1,7 @@
 package com.tenco.indiepicter.banner;
 
 import com.tenco.indiepicter.funding.FundingResponse;
+import com.tenco.indiepicter.funding.response.BannerDTO;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,10 +17,10 @@ public class BannerService {
     @Autowired
     private BannerRepository bannerRepository;
 
-    public List<FundingResponse.FundingPlusDTO.BannerDTO> DisplayBanner(String genre){
+    public List<BannerDTO> DisplayBanner(String genre){
         return bannerRepository.findByGenre(genre)
                 .stream()
-                .map(FundingResponse.FundingPlusDTO.BannerDTO::new)
+                .map(BannerDTO::new)
                 .collect(Collectors.toList());
     }
 }
