+++ src/main/java/com/tenco/indiepicter/funding/FundingRepository.java
@@ -1,6 +1,7 @@
 package com.tenco.indiepicter.funding;
 
 
+import com.tenco.indiepicter.funding.response.MoviesByGenreDTO;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
 
@@ -9,6 +10,6 @@ import java.util.List;
 @Mapper
 public interface FundingRepository {
     //펀딩 장르별 조회
-    public List<FundingResponse.FundingPlusDTO.MoviesByGenreDTO> findAllByGenre(@Param("genre") String genre, @Param("pageSize") Integer pageSize, @Param("offset") Integer offset);
+    public List<MoviesByGenreDTO> findAllByGenre(@Param("genre") String genre, @Param("pageSize") Integer pageSize, @Param("offset") Integer offset);
 
 }
